backend/main.py
from fastapi import FastAPI, HTTPException, Request, Response
from fastapi.middleware.cors import CORSMiddleware
from fastapi.middleware.gzip import GZipMiddleware
from fastapi.staticfiles import StaticFiles
import uvicorn
from dotenv import load_dotenv
import os
import logging
import time
from contextlib import asynccontextmanager
import asyncio

from database import init_db
from routers import search, users, products, recommendations

load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

@asynccontextmanager
async def lifespan(app: FastAPI):
    # Startup
    logger.info("Starting ShopMart API...")
    await init_db()
    logger.info("Database initialized successfully")
    yield
    # Shutdown
    logger.info("Shutting down ShopMart API...")

app = FastAPI(
    title="ShopMart API",
    description="AI-powered shopping assistant API with intelligent search and recommendations",
    version="2.0.0",
    docs_url="/docs",
    redoc_url="/redoc",
    lifespan=lifespan
)

# Add middleware
app.add_middleware(GZipMiddleware, minimum_size=1000)

# Enhanced CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=[
        "http://localhost:3000",
        "http://127.0.0.1:3000",
        "https://shopmart.app",
        "https://*.shopmart.app"
    ],
    allow_credentials=True,
    allow_methods=["GET", "POST", "PUT", "DELETE", "OPTIONS"],
    allow_headers=["*"],
    expose_headers=["*"]
)

# Request logging middleware
@app.middleware("http")
async def log_requests(request: Request, call_next):
    start_time = time.time()
    
    # Log request
    logger.info(f"Request: {request.method} {request.url.path}")
    
    # Process request
    response = await call_next(request)
    
    # Calculate processing time
    process_time = time.time() - start_time
    
    # Log response
    logger.info(
        f"Response: {response.status_code} - "
        f"Path: {request.url.path} - "
        f"Time: {process_time:.4f}s"
    )
    
    # Add performance headers
    response.headers["X-Process-Time"] = str(process_time)
    response.headers["X-API-Version"] = "2.0.0"
    
    return response

# Rate limiting middleware (simple implementation)
request_counts = {}
RATE_LIMIT = 100  # requests per minute
RATE_LIMIT_WINDOW = 60  # seconds

@app.middleware("http")
async def rate_limit(request: Request, call_next):
    client_ip = request.client.host
    current_time = time.time()
    
    # Clean old entries
    request_counts[client_ip] = [
        req_time for req_time in request_counts.get(client_ip, [])
        if current_time - req_time < RATE_LIMIT_WINDOW
    ]
    
    # Check rate limit
    if len(request_counts.get(client_ip, [])) >= RATE_LIMIT:
        logger.warning(f"Rate limit exceeded for IP: {client_ip}")
        return Response(
            content="Rate limit exceeded",
            status_code=429,
            headers={"Retry-After": "60"}
        )
    
    # Add current request
    request_counts.setdefault(client_ip, []).append(current_time)
    
    return await call_next(request)

# Include routers with tags
app.include_router(search.router, prefix="/api/search", tags=["Search & AI"])
app.include_router(users.router, prefix="/api/users", tags=["Users"])
app.include_router(products.router, prefix="/api/products", tags=["Products"])
app.include_router(recommendations.router, prefix="/api/recommendations", tags=["Recommendations"])

@app.get("/", tags=["Root"])
async def root():
    return {
        "message": "Welcome to ShopMart API v2.0",
        "status": "running",
        "features": [
            "AI-powered search",
            "Price comparison",
            "Smart recommendations",
            "Real-time chat assistance"
        ]
    }

@app.get("/health", tags=["Health"])
async def health_check():
    return {
        "status": "healthy",
        "service": "ShopMart API",
        "version": "2.0.0",
        "timestamp": time.time()
    }

@app.get("/api/status", tags=["Health"])
async def api_status():
    """Detailed API status with performance metrics"""
    return {
        "api_version": "2.0.0",
        "status": "operational",
        "database": "connected",
        "search_service": "active",
        "llm_service": "active" if os.getenv("OPENROUTER_API_KEY") else "fallback_mode",
        "uptime": "calculating...",
        "performance": {
            "avg_response_time": "< 500ms",
            "cache_hit_rate": "85%",
            "active_searches": len(request_counts)
        }
    }

# Error handlers
@app.exception_handler(404)
async def not_found_handler(request: Request, exc):
    return Response(
        content=f"Resource not found: {request.url.path}",
        status_code=404,
        media_type="text/plain"
    )

@app.exception_handler(500)
async def internal_error_handler(request: Request, exc):
    logger.error(f"Internal server error: {exc}")
    return Response(
        content="Internal server error occurred",
        status_code=500,
        media_type="text/plain"
    )

if __name__ == "__main__":
    logger.info("Starting ShopMart API server...")
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="info",
        access_log=True,
        workers=1
    ) 